Revert "Internal: Cherry-pick PR 32949 to 3.32 Update core onboarding flow [E…"
This reverts commit 26e5f5923042ce18f72c6977032ed5abcc70a012.
