--- app/modules/onboarding/assets/js/components/theme-selection-content-b.js
@@ -1,221 +1,0 @@
-import PropTypes from 'prop-types';
-import { useState, useCallback } from 'react';
-import PageContentLayout from './layout/page-content-layout';
-import ThemeSelectionCard from './theme-selection-card';
-
-const getThemeData = () => [
-	{
-		slug: 'hello-theme',
-		title: __( 'Hello', 'elementor' ),
-		description: __( 'It\'s fast, flexible, and beginner-friendly, offering a solid foundation for customizable designs.', 'elementor' ),
-		illustration: {
-			svg: (
-				<svg width="305" height="164" viewBox="0 0 305 164" fill="none" xmlns="http://www.w3.org/2000/svg">
-					<rect x="97.5146" y="130.5" width="0.398535" height="0.398535" transform="rotate(-2.18781 97.5146 130.5)" fill="#cccccc" />
-					<path d="M16.1259 137.899L14.1372 138.024L19.5181 223.129L21.5068 223.004L16.1259 137.899Z" fill="black" />
-					<path d="M9.44682 233.715L14.0914 19.0745L212.074 19.1003L207.454 233.751L9.44682 233.715Z" fill="white" />
-					<path d="M9.44682 233.715L14.0914 19.0745L212.074 19.1003L207.454 233.751L9.44682 233.715Z" fill="white" />
-					<path d="M202.793 233.962L207.437 19.3215L219.305 19.5368L214.672 234.188L202.793 233.962Z" fill="#F0F0FF" />
-					<path d="M208.415 20.3405L218.287 20.5147L213.696 233.17L203.718 232.989L208.31 20.3445L208.415 20.3405ZM206.367 18.319L201.679 234.896L215.648 235.16L220.275 18.5267L206.367 18.319Z" fill="black" />
-					<rect x="103.898" y="130.257" width="0.398535" height="0.398535" transform="rotate(-2.18781 103.898 130.257)" fill="#cccccc" />
-					<path d="M22.8358 87.0786L22.9487 86.042L23.2435 87.0044L22.8358 87.0786Z" fill="black" />
-					<path d="M23.2303 79.285L23.2173 78.9453C23.2173 78.9453 23.3276 79.0701 23.329 79.1053C23.3303 79.1404 23.2632 79.2251 23.2303 79.285Z" fill="black" />
-					<path d="M75.7296 51.8251C79.0103 51.4182 81.0516 52.654 81.113 55.4904L72.8915 55.218C72.8158 54.8197 72.8209 54.4104 72.9065 54.0141C72.9921 53.6179 73.1565 53.243 73.3899 52.9115C73.6798 52.574 74.0386 52.3024 74.4421 52.115C74.8457 51.9276 75.2847 51.8288 75.7296 51.8251Z" fill="white" />
-					<path d="M206.102 83.8532C206.837 83.8597 207.565 84.011 208.242 84.2987C208.919 84.5863 209.533 85.0045 210.049 85.5294C210.565 86.0544 210.972 86.6757 211.247 87.358C211.523 88.0404 211.661 88.7703 211.655 89.5061L211.405 121.035C211.398 121.77 211.247 122.498 210.959 123.175C210.672 123.852 210.253 124.466 209.728 124.982C209.204 125.498 208.582 125.905 207.9 126.18C207.218 126.456 206.488 126.594 205.752 126.588C205.016 126.583 204.287 126.433 203.609 126.145C202.931 125.858 202.317 125.44 201.801 124.915C201.285 124.39 200.878 123.768 200.603 123.085C200.328 122.402 200.19 121.671 200.199 120.935L200.46 89.4059C200.474 87.922 201.075 86.5039 202.133 85.4629C203.19 84.4219 204.618 83.843 206.102 83.8532Z" fill="white" />
-					<mask id="mask0_222_51325" style={ { maskType: 'alpha' } } maskUnits="userSpaceOnUse" x="131" y="42" width="212" height="172">
-						<path d="M173.627 46.9854C173.627 46.9854 108.276 62.8966 141.236 154.779C174.197 246.662 296.933 204.108 296.933 204.108C296.933 204.108 378.131 184.124 324.212 99.8914C273.072 20.0029 173.627 46.9854 173.627 46.9854Z" fill="#F0F0FF" />
-					</mask>
-					<g mask="url(#mask0_222_51325)">
-						<path d="M206.169 18.7754L14.5165 18.7754L10.0801 234.106L201.288 234.106L206.169 18.7754Z" fill="#DCDBFF" />
-					</g>
-					<rect x="199.67" y="83.749" width="12.0793" height="43.2336" rx="6.03966" fill="white" stroke="black" strokeWidth="2" />
-					<path d="M36.6708 86.7568C36.9784 86.8337 37.2592 86.9072 37.5801 86.9306C39.793 87.1045 43.9213 87.1312 46.1142 86.9306C46.5487 86.8905 46.8897 86.8671 47.1905 86.5061V77.8249L47.4412 77.5742H52.2849C52.3417 77.5742 52.539 77.7948 52.7028 77.7414V101.117H47.1905V91.8512L46.9398 91.6005H36.6675V101.117H31.1553V77.8249L31.406 77.5742H36.4168L36.6675 77.8249V86.7568H36.6708Z" fill="black" />
-					<path d="M75.253 93.6032H62.1426C61.3738 93.6032 62.5605 95.2846 62.7544 95.4952C64.2854 97.1599 67.3039 97.3404 69.3028 96.585C70.1085 96.2808 70.5932 95.6757 71.3219 95.4284L71.6762 95.5922L74.0696 98.5004C73.1403 99.6302 71.7364 100.539 70.319 100.937C64.5494 102.555 57.0081 100.088 56.3864 93.3458C55.0627 79.0221 77.3355 79.3664 75.2563 93.5999L75.253 93.6032ZM61.9755 90.4309H69.9915C70.4762 90.4309 69.9513 89.2977 69.8343 89.0871C68.3368 86.4029 64.4759 86.1355 62.5571 88.5088C62.2931 88.8364 61.3471 90.2571 61.9755 90.4343V90.4309Z" fill="black" />
-					<path d="M105.946 82.7993C113.083 82.0639 118.826 87.4825 116.917 94.7697C114.283 104.805 96.1218 103.605 97.4856 90.5378C97.9235 86.3493 101.915 83.2171 105.95 82.8026L105.946 82.7993ZM106.441 87.2986C101.654 88.0474 101.597 96.2405 106.611 96.8991C113.604 97.8183 113.534 86.1888 106.441 87.2986Z" fill="black" />
-					<path d="M78.843 76.2373H83.8538L84.1045 76.488V100.867L83.8538 101.118H78.843L78.5923 100.867V76.488L78.843 76.2373Z" fill="black" />
-					<path d="M88.8616 76.2373H93.8724L94.1231 76.488V100.867L93.8724 101.118H88.8616L88.6143 100.867V76.488L88.8616 76.2373Z" fill="black" />
-					<path d="M49.8644 111.634C49.3229 111.57 48.6008 111.848 48.1997 111.383C48.7045 110.648 49.8811 110.564 49.8644 111.634Z" fill="white" />
-					<path fillRule="evenodd" clipRule="evenodd" d="M32.9312 62.6564C31.7733 60.9235 31.1553 58.8862 31.1553 56.8021C31.1553 54.0074 32.2655 51.3272 34.2416 49.351C36.2178 47.3748 38.898 46.2646 41.6927 46.2646C43.7768 46.2646 45.8142 46.8827 47.547 48.0405C49.2799 49.1984 50.6305 50.8441 51.4281 52.7696C52.2256 54.6951 52.4343 56.8138 52.0277 58.8579C51.6211 60.9019 50.6176 62.7795 49.1438 64.2532C47.6702 65.7269 45.7926 66.7305 43.7485 67.1371C41.7044 67.5437 39.5857 67.335 37.6602 66.5374C35.7348 65.7399 34.089 64.3893 32.9312 62.6564ZM39.0588 52.4112H37.3028V61.1926H39.0588V52.4112ZM46.0827 52.4112H40.8148V54.1671H46.0827V52.4112ZM46.0827 55.9231H40.8148V57.6791H46.0827V55.9231ZM46.0827 59.4367H40.8148V61.1926H46.0827V59.4367Z" fill="black" />
-					<path d="M122.785 115.106V115.565H31.1553V115.106H122.785Z" fill="black" />
-					<path d="M122.785 121.979V122.437H31.1553V121.979H122.785Z" fill="black" />
-					<path d="M122.785 128.851V129.309H31.1553V128.851H122.785Z" fill="black" />
-					<path d="M31.1553 143.759C31.1553 140.216 34.027 137.345 37.5694 137.345H80.1774C83.7198 137.345 86.5915 140.216 86.5915 143.759C86.5915 147.301 83.7198 150.173 80.1774 150.173H37.5694C34.027 150.173 31.1553 147.301 31.1553 143.759Z" fill="black" />
-					<path d="M15.4765 14.9994L219.902 14.9155L214.999 233.424L10.5366 233.509L15.4765 14.9994ZM13.4435 13.0282L8.48565 235.552L217.096 235.465L222.051 12.8688L13.4398 12.9319L13.4435 13.0282Z" fill="black" />
-					<path d="M14.6611 22.8996V12.8203L221.745 13.4923V22.8996H14.6611Z" fill="black" />
-				</svg>
-			),
-		},
-	},
-	{
-		slug: 'hello-biz',
-		title: __( 'Hello Biz', 'elementor' ),
-		description: __( 'It offers premium Elementor tools, and a responsive foundation for startups and portfolios.', 'elementor' ),
-		illustration: {
-			svg: (
-				<svg width="231" height="170" viewBox="0 0 231 170" fill="none" xmlns="http://www.w3.org/2000/svg">
-					<rect x="97.5146" y="125.526" width="0.398535" height="0.398535" transform="rotate(-2.18781 97.5146 125.526)" fill="#cccccc" />
-					<path d="M16.1264 132.926L14.1377 133.052L19.5186 218.157L21.5073 218.031L16.1264 132.926Z" fill="black" />
-					<path d="M9.44682 228.741L14.0914 14.1008L212.074 14.1267L207.454 228.778L9.44682 228.741Z" fill="white" />
-					<path d="M9.44682 228.741L14.0914 14.1008L212.074 14.1267L207.454 228.778L9.44682 228.741Z" fill="white" />
-					<path d="M202.793 228.988L207.437 14.3479L219.305 14.5632L214.672 229.215L202.793 228.988Z" fill="#F0F0FF" />
-					<rect x="103.898" y="125.283" width="0.398535" height="0.398535" transform="rotate(-2.18781 103.898 125.283)" fill="#cccccc" />
-					<path d="M22.8358 82.105L22.9487 81.0684L23.2435 82.0307L22.8358 82.105Z" fill="black" />
-					<path d="M23.2303 74.3114L23.2173 73.9717C23.2173 73.9717 23.3276 74.0965 23.329 74.1316C23.3303 74.1668 23.2632 74.2515 23.2303 74.3114Z" fill="black" />
-					<path d="M75.7296 46.8515C79.0103 46.4446 81.0516 47.6804 81.113 50.5168L72.8915 50.2443C72.8158 49.8461 72.8209 49.4367 72.9065 49.0405C72.9921 48.6443 73.1565 48.2693 73.3899 47.9379C73.6798 47.6004 74.0386 47.3288 74.4421 47.1414C74.8457 46.954 75.2847 46.8551 75.7296 46.8515Z" fill="white" />
-					<path d="M206.102 78.8795C206.837 78.886 207.565 79.0374 208.242 79.325C208.919 79.6126 209.533 80.0309 210.049 80.5558C210.565 81.0807 210.972 81.7021 211.247 82.3844C211.523 83.0667 211.661 83.7966 211.655 84.5325L211.405 116.061C211.398 116.797 211.247 117.524 210.959 118.201C210.672 118.879 210.253 119.493 209.728 120.008C209.204 120.524 208.582 120.931 207.9 121.207C207.218 121.482 206.488 121.621 205.752 121.614C205.016 121.609 204.287 121.459 203.609 121.172C202.931 120.885 202.317 120.467 201.801 119.941C201.285 119.416 200.878 118.794 200.603 118.111C200.328 117.428 200.19 116.697 200.199 115.961L200.46 84.4323C200.474 82.9483 201.075 81.5303 202.133 80.4893C203.19 79.4483 204.618 78.8693 206.102 78.8795Z" fill="white" />
-					<rect x="199.67" y="78.7754" width="12.0793" height="43.2336" rx="6.03966" fill="white" stroke="black" strokeWidth="2" />
-					<g clipPath="url(#clip0_222_51393)">
-						<path d="M148.416 69.1784C147.281 70.2262 147.566 73.0909 149.384 73.5034C154.284 74.618 154.125 68.5235 150.444 68.543C150.03 68.543 148.693 68.9248 148.419 69.1784H148.416Z" fill="white" />
-						<path d="M136.46 69.3849C133.193 69.0199 129.309 69.6552 125.972 69.3849V94.1866H136.46C137.095 94.1866 139.629 93.568 140.317 93.2921C145.046 91.3916 145.796 83.4912 140.795 81.6436V81.0947C143.074 80.3005 143.262 76.9007 143.033 74.819C142.675 71.5836 139.461 69.7221 136.46 69.3849Z" fill="white" />
-						<path d="M36.755 69.6631H32.0005V93.9075H36.755V83.8753H48.2218V93.9075H52.9763V69.6631H48.2218V79.6952H37.1745L36.755 79.2772V69.6631Z" fill="white" />
-						<path d="M82.902 69.6631H75.91L75.9072 73.0099L78.1474 73.0071V93.9075H82.902V69.6631Z" fill="white" />
-						<path d="M92.4108 69.6631H85.4188L85.416 73.0099L87.6562 73.0071V93.9075H92.4108V69.6631Z" fill="white" />
-						<path d="M61.0871 86.1052H74.3718C75.0402 85.5311 74.8332 84.3217 74.7549 83.4912C74.1928 77.4051 69.0299 74.1 63.1427 75.4739C55.4012 77.2825 54.33 89.5412 61.0004 93.2949C65.7549 95.9729 73.4097 94.3427 74.2319 88.3318H70.0367C69.9892 88.3318 69.9584 88.8473 69.634 89.1733C66.6331 92.183 61.3472 90.55 61.0871 86.1024V86.1052Z" fill="white" />
-						<path d="M103.531 75.3088C93.0484 76.8165 93.1882 93.1939 103.768 94.434C119.883 96.3234 119.33 73.0349 103.531 75.3088Z" fill="white" />
-						<path d="M153.381 75.7939H146.109C146.146 78.1013 145.385 80.3586 148.626 79.138V93.9076H153.381V75.7939Z" fill="white" />
-						<path d="M170.721 75.7939H157.156C156.488 76.5519 156.477 78.628 157.156 79.4167H165.267L165.55 79.974L157.229 89.7999L157.017 93.9076H171V90.2848H162.051C162.054 89.7163 162.565 89.198 162.895 88.7577C165.217 85.6673 168.19 82.7663 170.575 79.6898L170.721 75.7939Z" fill="white" />
-						<path d="M136.46 69.3849C139.464 69.7221 142.678 71.5836 143.033 74.819C143.262 76.9007 143.075 80.3005 140.795 81.0947V81.6436C145.793 83.494 145.047 91.3943 140.314 93.2921C139.626 93.568 137.093 94.1866 136.458 94.1866H125.97V69.3849C129.306 69.6552 133.191 69.0199 136.458 69.3849H136.46ZM130.447 79.6957H136.46C137.834 79.6957 138.337 77.2852 138.292 76.204C138.25 75.237 137.568 73.2863 136.46 73.2863H130.447V79.6957ZM130.447 90.2852H137.02C137.82 90.2852 139.162 88.694 139.333 87.8524C140.488 82.1703 134.41 83.9538 131.018 83.4327L130.447 83.7365V90.2852Z" fill="black" />
-						<path d="M36.755 69.6631V79.2772L37.1745 79.6952H48.2218V69.6631H52.9763V93.9075H48.2218V83.8753H36.755V93.9075H32.0005V69.6631H36.755Z" fill="black" />
-						<path d="M61.0871 86.1043C61.3472 90.5519 66.6331 92.1849 69.634 89.1753C69.9584 88.852 69.9892 88.3337 70.0367 88.3337H74.2319C73.4097 94.3446 65.7577 95.9721 61.0004 93.2968C54.33 89.5431 55.4012 77.2816 63.1427 75.4758C69.0299 74.102 74.1928 77.407 74.7549 83.4932C74.8304 84.3236 75.0374 85.5331 74.3718 86.1071H61.0871V86.1043ZM70.3164 83.039C69.4886 77.538 62.1638 77.7804 61.0871 83.039H70.3164Z" fill="black" />
-						<path d="M103.531 75.3088C119.327 73.0349 119.884 96.3262 103.769 94.434C93.1914 93.1939 93.0515 76.8164 103.531 75.3088ZM104.641 79.2046C98.9136 80.1633 99.0842 89.7886 104.599 90.5382C113.32 91.7225 112.825 77.8364 104.641 79.2046Z" fill="black" />
-						<path d="M170.721 75.7939L170.575 79.6898C168.19 82.7663 165.217 85.67 162.895 88.7577C162.565 89.198 162.054 89.7163 162.051 90.2848H171V93.9076H157.017L157.229 89.7999L165.55 79.974L165.267 79.4167H157.156C156.477 78.628 156.488 76.5519 157.156 75.7939H170.721Z" fill="black" />
-						<path d="M92.4108 69.6631V93.9075H87.6562V73.0071L85.416 73.0099L85.4188 69.6631H92.4108Z" fill="black" />
-						<path d="M82.902 69.6631V93.9075H78.1474V73.0071L75.9072 73.0099L75.91 69.6631H82.902Z" fill="black" />
-						<path d="M153.381 75.7939V93.9076H148.626V79.138C145.385 80.3586 146.146 78.1013 146.109 75.7939H153.381Z" fill="black" />
-						<path d="M148.416 69.1784C148.693 68.9248 150.027 68.5458 150.441 68.543C154.122 68.5235 154.281 74.618 149.381 73.5034C147.566 73.0909 147.278 70.2234 148.414 69.1784H148.416Z" fill="black" />
-						<path d="M130.447 90.2851V83.7364L131.018 83.4326C134.408 83.9537 140.488 82.1702 139.333 87.8523C139.162 88.6939 137.82 90.2851 137.02 90.2851H130.447Z" fill="white" />
-						<path d="M130.447 79.6956V73.2861H136.46C137.568 73.2861 138.25 75.234 138.292 76.2038C138.337 77.2878 137.834 79.6956 136.46 79.6956H130.447Z" fill="white" />
-						<path d="M70.3163 83.0392H61.0869C62.1637 77.7807 69.4884 77.5383 70.3163 83.0392Z" fill="white" />
-						<path d="M104.641 79.2054C112.824 77.8371 113.319 91.7205 104.599 90.5389C99.0839 89.7893 98.9133 80.164 104.641 79.2054Z" fill="white" />
-					</g>
-					<path fillRule="evenodd" clipRule="evenodd" d="M32.9307 57.6828C31.7728 55.9499 31.1548 53.9126 31.1548 51.8285C31.1548 49.0338 32.265 46.3535 34.2411 44.3774C36.2173 42.4012 38.8976 41.291 41.6922 41.291C43.7764 41.291 45.8137 41.909 47.5465 43.0669C49.2794 44.2248 50.63 45.8705 51.4276 47.796C52.2251 49.7214 52.4338 51.8402 52.0272 53.8842C51.6206 55.9283 50.6171 57.8059 49.1433 59.2796C47.6697 60.7533 45.7921 61.7569 43.748 62.1635C41.7039 62.5701 39.5852 62.3614 37.6597 61.5638C35.7343 60.7663 34.0885 59.4157 32.9307 57.6828ZM39.0583 47.4375H37.3023V56.219H39.0583V47.4375ZM46.0822 47.4375H40.8143V49.1935H46.0822V47.4375ZM46.0822 50.9495H40.8143V52.7055H46.0822V50.9495ZM46.0822 54.463H40.8143V56.219H46.0822V54.463Z" fill="black" />
-					<path d="M122.785 107.133V107.591H31.1548V107.133H122.785Z" fill="black" />
-					<path d="M122.785 114.877V115.335H31.1548V114.877H122.785Z" fill="black" />
-					<path d="M31.1548 130.785C31.1548 127.243 34.0265 124.371 37.5689 124.371H80.1769C83.7193 124.371 86.591 127.243 86.591 130.785C86.591 134.328 83.7193 137.199 80.1769 137.199H37.5689C34.0265 137.199 31.1548 134.328 31.1548 130.785Z" fill="black" />
-					<path d="M15.4765 10.0258L219.902 9.94184L214.999 228.45L10.5366 228.535L15.4765 10.0258ZM13.4435 8.05454L8.48565 230.578L217.096 230.491L222.051 7.89518L13.4398 7.95826L13.4435 8.05454Z" fill="black" />
-					<path d="M14.6611 17.926V7.84668L221.745 8.51863V17.926H14.6611Z" fill="black" />
-					<rect x="29.0005" y="148.543" width="38" height="38" rx="6" fill="#DCDBFF" />
-					<rect x="71.0005" y="148.543" width="38" height="38" rx="6" fill="#DCDBFF" />
-					<rect x="114" y="148.543" width="38" height="38" rx="6" fill="#DCDBFF" />
-					<rect x="156" y="148.543" width="38" height="38" rx="6" fill="#DCDBFF" />
-				</svg>
-			),
-			className: 'e-onboarding__theme-card-illustration--hello-biz',
-		},
-	},
-];
-
-export default function ThemeSelectionContentB( { actionButton, skipButton, noticeState, selectedTheme, onThemeSelect, onThemeInstallSuccess, onThemeInstallError } ) {
-	const [ isInstalling, setIsInstalling ] = useState( false );
-	const [ installingTheme, setInstallingTheme ] = useState( null );
-
-	const resetInstallationState = useCallback( () => {
-		setIsInstalling( false );
-		setInstallingTheme( null );
-	}, [] );
-
-	const activateTheme = useCallback( ( themeSlug ) => {
-		const themeSlugForAPI = 'hello-theme' === themeSlug ? 'hello-elementor' : 'hello-biz';
-
-		const formData = new FormData();
-		formData.append( '_nonce', elementorCommon.config.ajax.nonce );
-		formData.append( 'action', 'elementor_activate_hello_theme' );
-		formData.append( 'theme_slug', themeSlugForAPI );
-
-		fetch( elementorCommon.config.ajax.url, {
-			method: 'POST',
-			body: formData,
-		} )
-			.then( ( response ) => response.json() )
-			.then( ( data ) => {
-				resetInstallationState();
-				if ( data.success && onThemeInstallSuccess ) {
-					onThemeInstallSuccess();
-				} else if ( onThemeInstallError ) {
-					onThemeInstallError();
-				}
-			} )
-			.catch( () => {
-				resetInstallationState();
-				if ( onThemeInstallError ) {
-					onThemeInstallError();
-				}
-			} );
-	}, [ onThemeInstallSuccess, onThemeInstallError, resetInstallationState ] );
-
-	const installTheme = useCallback( ( themeSlug ) => {
-		const themeSlugForAPI = 'hello-theme' === themeSlug ? 'hello-elementor' : 'hello-biz';
-		wp.updates.ajax( 'install-theme', {
-			slug: themeSlugForAPI,
-			success: () => activateTheme( themeSlug ),
-			error: () => {
-				resetInstallationState();
-				if ( onThemeInstallError ) {
-					onThemeInstallError();
-				}
-			},
-		} );
-	}, [ activateTheme, onThemeInstallError, resetInstallationState ] );
-
-	const startThemeInstallation = useCallback( ( themeSlug ) => {
-		setIsInstalling( true );
-		setInstallingTheme( themeSlug );
-		installTheme( themeSlug );
-	}, [ installTheme ] );
-
-	const handleThemeSelect = useCallback( ( themeSlug ) => {
-		if ( onThemeSelect ) {
-			onThemeSelect( themeSlug );
-		}
-
-		startThemeInstallation( themeSlug );
-	}, [ onThemeSelect, startThemeInstallation ] );
-
-	const themeData = getThemeData();
-
-	return (
-		<PageContentLayout
-			title={ __( 'Which theme would you like?', 'elementor' ) }
-			actionButton={ actionButton }
-			skipButton={ skipButton }
-			noticeState={ noticeState }
-		>
-			<div className="e-onboarding__page-content-theme-variant-b">
-				<p className="e-onboarding__theme-selection-description">
-					{ __( 'The theme delivers fast setup, intuitive tools, and business-focused widgets.', 'elementor' ) }
-				</p>
-				<p className="e-onboarding__theme-selection-subtitle">
-					{ __( 'You can switch your theme later on', 'elementor' ) }
-				</p>
-
-				<div className="e-onboarding__theme-cards">
-					{ themeData.map( ( theme ) => (
-						<ThemeSelectionCard
-							key={ theme.slug }
-							themeSlug={ theme.slug }
-							title={ theme.title }
-							description={ theme.description }
-							illustration={ theme.illustration }
-							isSelected={ selectedTheme === theme.slug }
-							isLoading={ isInstalling && installingTheme === theme.slug }
-							onSelect={ handleThemeSelect }
-							aria-label={ `Select ${ theme.title } theme: ${ theme.description }` }
-							role="button"
-							tabIndex={ 0 }
-						/>
-					) ) }
-				</div>
-			</div>
-		</PageContentLayout>
-	);
-}
-
-ThemeSelectionContentB.propTypes = {
-	actionButton: PropTypes.object.isRequired,
-	skipButton: PropTypes.object.isRequired,
-	noticeState: PropTypes.object,
-	selectedTheme: PropTypes.string,
-	onThemeSelect: PropTypes.func,
-	onThemeInstallSuccess: PropTypes.func,
-	onThemeInstallError: PropTypes.func,
-};
