+++ app/modules/onboarding/assets/js/pages/hello-theme.js
@@ -4,9 +4,8 @@ import { OnboardingContext } from '../context/context';
 import { useNavigate } from '@reach/router';
 import useAjax from 'elementor-app/hooks/use-ajax';
 import Layout from '../components/layout/layout';
-import ThemeSelectionContentA from '../components/theme-selection-content-a';
-import ThemeSelectionContentB from '../components/theme-selection-content-b';
-import { OnboardingEventTracking, ONBOARDING_STORAGE_KEYS } from '../utils/onboarding-event-tracking';
+import PageContentLayout from '../components/layout/page-content-layout';
+import { OnboardingEventTracking } from '../utils/onboarding-event-tracking';
 
 export default function HelloTheme() {
 	const { state, updateState, getStateObjectToUpdate } = useContext( OnboardingContext ),
@@ -14,7 +13,6 @@ export default function HelloTheme() {
 		// Allow navigating back to this screen if it was completed in the onboarding.
 		[ helloInstalledInOnboarding, setHelloInstalledInOnboarding ] = useState( false ),
 		[ isInstalling, setIsInstalling ] = useState( false ),
-		[ selectedTheme, setSelectedTheme ] = useState( null ),
 		noticeStateSuccess = {
 			type: 'success',
 			icon: 'eicon-check-circle-o',
@@ -46,7 +44,7 @@ export default function HelloTheme() {
 
 		OnboardingEventTracking.setupAllUpgradeButtons( state.currentStep );
 		OnboardingEventTracking.onStepLoad( 2 );
-	}, [ getStateObjectToUpdate, goToNextScreen, helloInstalledInOnboarding, pageId, state, updateState ] );
+	}, [] );
 
 	const resetScreenContent = () => {
 		// Clear any active timeouts for changing the action button text during installation.
@@ -90,7 +88,7 @@ export default function HelloTheme() {
 
 		OnboardingEventTracking.sendStepEndState( 2 );
 		goToNextScreen();
-	}, [ getStateObjectToUpdate, goToNextScreen, noticeStateSuccess, state, updateState ] );
+	}, [] );
 
 	const onErrorInstallHelloTheme = () => {
 		elementorCommon.events.dispatchEvent( {
@@ -118,13 +116,8 @@ export default function HelloTheme() {
 
 		updateState( { isHelloThemeInstalled: true } );
 
-		const themeSlug = 'hello-theme' === selectedTheme ? 'hello-elementor' : 'hello-biz';
-
 		setActivateHelloThemeAjaxState( {
-			data: {
-				action: 'elementor_activate_hello_theme',
-				theme_slug: themeSlug,
-			},
+			data: { action: 'elementor_activate_hello_theme' },
 		} );
 	};
 
@@ -133,10 +126,8 @@ export default function HelloTheme() {
 			setIsInstalling( true );
 		}
 
-		const themeSlug = 'hello-theme' === selectedTheme ? 'hello-elementor' : 'hello-biz';
-
 		wp.updates.ajax( 'install-theme', {
-			slug: themeSlug,
+			slug: 'hello-biz',
 			success: () => activateHelloTheme(),
 			error: () => onErrorInstallHelloTheme(),
 		} );
@@ -151,15 +142,6 @@ export default function HelloTheme() {
 				step: state.currentStep,
 			},
 		} );
-	};
-
-	const handleThemeSelection = ( themeSlug ) => {
-		setSelectedTheme( themeSlug );
-
-		const themeValue = 'hello-theme' === themeSlug ? 'hello' : 'hellobiz';
-
-		OnboardingEventTracking.trackStepAction( 2, `select_theme_${ themeSlug.replace( '-', '_' ) }`, { theme: themeValue } );
-		OnboardingEventTracking.sendThemeChoiceEvent( state.currentStep, themeValue );
 	};
 
 	/**
@@ -287,20 +269,27 @@ export default function HelloTheme() {
 		}
 	}, [ activateHelloThemeAjaxState.status ] );
 
-	const variant = localStorage.getItem( ONBOARDING_STORAGE_KEYS.THEME_SELECTION_VARIANT );
-	const ContentComponent = 'B' === variant ? ThemeSelectionContentB : ThemeSelectionContentA;
-
 	return (
 		<Layout pageId={ pageId } nextStep={ nextStep }>
-			<ContentComponent
+			<PageContentLayout
+				image={ elementorCommon.config.urls.assets + 'images/app/onboarding/Illustration_Hello_Biz.svg' }
+				title={ __( 'Every site starts with a theme.', 'elementor' ) }
 				actionButton={ actionButton }
 				skipButton={ skipButton }
 				noticeState={ noticeState }
-				selectedTheme={ selectedTheme }
-				onThemeSelect={ handleThemeSelection }
-				onThemeInstallSuccess={ onHelloThemeActivationSuccess }
-				onThemeInstallError={ onErrorInstallHelloTheme }
-			/>
+			>
+				<p>
+					{ __( 'Hello Biz by Elementor helps you launch your professional business website - fast.', 'elementor' ) }
+				</p>
+				{ ! elementorAppConfig.onboarding.experiment && <p>
+					{ __( 'Here\'s why:', 'elementor' ) }
+				</p> }
+				<ul className="e-onboarding__feature-list">
+					<li>{ __( 'Get online faster', 'elementor' ) }</li>
+					<li>{ __( 'Lightweight and fast loading', 'elementor' ) }</li>
+					<li>{ __( 'Great for SEO', 'elementor' ) }</li>
+				</ul>
+			</PageContentLayout>
 			<div className="e-onboarding__footnote">
 				{ '* ' + __( 'You can switch your theme later on', 'elementor' ) }
 			</div>
