+++ app/modules/onboarding/assets/js/utils/modules/click-tracker.js
@@ -369,7 +369,6 @@ class ClickTracker {
 		}
 		this.lastSelectClickTime = 0;
 		StorageManager.clearAllOnboardingData();
-		StorageManager.clearExperimentData();
 	}
 }
