+++ app/modules/onboarding/assets/js/utils/modules/event-dispatcher.js
@@ -1,10 +1,6 @@
-import StorageManager, { ONBOARDING_STORAGE_KEYS } from './storage-manager.js';
-
 export const ONBOARDING_EVENTS_MAP = {
 	UPGRADE_NOW_S3: 'core_onboarding_s3_upgrade_now',
 	HELLO_BIZ_CONTINUE: 'core_onboarding_s2_hellobiz',
-	THEME_CHOICE: 'core_onboarding_theme_choice',
-	THEME_CHOICE_VARIANT_B: 'ab_201_theme_choice',
 	CORE_ONBOARDING: 'core_onboarding',
 	CONNECT_STATUS: 'core_onboarding_connect_status',
 	STEP1_END_STATE: 'core_onboarding_s1_end_state',
@@ -67,28 +63,6 @@ export function dispatchIfAllowed( eventName, payload = {} ) {
 	return false;
 }
 
-function getThemeSelectionVariant() {
-	const variant = StorageManager.getString( ONBOARDING_STORAGE_KEYS.THEME_SELECTION_VARIANT );
-	return variant || null;
-}
-
-function getGoodToGoVariant() {
-	const variant = StorageManager.getString( ONBOARDING_STORAGE_KEYS.GOOD_TO_GO_VARIANT );
-	return variant || null;
-}
-
-function getVariantSpecificEventName( baseEventName, stepNumber ) {
-	if ( 2 === stepNumber && elementorAppConfig?.onboarding?.themeSelectionExperimentEnabled ) {
-		const variant = getThemeSelectionVariant();
-		if ( 'B' === variant ) {
-			if ( ONBOARDING_EVENTS_MAP.THEME_CHOICE === baseEventName ) {
-				return ONBOARDING_EVENTS_MAP.THEME_CHOICE_VARIANT_B;
-			}
-		}
-	}
-	return baseEventName;
-}
-
 export function createEventPayload( basePayload = {} ) {
 	return {
 		location: 'plugin_onboarding',
@@ -98,21 +72,11 @@ export function createEventPayload( basePayload = {} ) {
 }
 
 export function createStepEventPayload( stepNumber, stepName, additionalData = {} ) {
-	const basePayload = {
+	return createEventPayload( {
 		step_number: stepNumber,
 		step_name: stepName,
 		...additionalData,
-	};
-
-	if ( 2 === stepNumber && elementorAppConfig?.onboarding?.themeSelectionExperimentEnabled ) {
-		basePayload.theme_selection_variant = getThemeSelectionVariant();
-	}
-
-	if ( 4 === stepNumber && elementorAppConfig?.onboarding?.goodToGoExperimentEnabled ) {
-		basePayload.good_to_go_variant = getGoodToGoVariant();
-	}
-
-	return createEventPayload( basePayload );
+	} );
 }
 
 export function createEditorEventPayload( additionalData = {} ) {
@@ -125,13 +89,7 @@ export function createEditorEventPayload( additionalData = {} ) {
 
 export function dispatchStepEvent( eventName, stepNumber, stepName, additionalData = {} ) {
 	const payload = createStepEventPayload( stepNumber, stepName, additionalData );
-	const variantSpecificEventName = getVariantSpecificEventName( eventName, stepNumber );
-	return dispatch( variantSpecificEventName, payload );
-}
-
-export function dispatchVariantAwareEvent( eventName, payload, stepNumber = null ) {
-	const variantSpecificEventName = stepNumber ? getVariantSpecificEventName( eventName, stepNumber ) : eventName;
-	return dispatch( variantSpecificEventName, payload );
+	return dispatch( eventName, payload );
 }
 
 export function dispatchEditorEvent( eventName, additionalData = {} ) {
@@ -192,7 +150,6 @@ const EventDispatcher = {
 	createStepEventPayload,
 	createEditorEventPayload,
 	dispatchStepEvent,
-	dispatchVariantAwareEvent,
 	dispatchEditorEvent,
 	getClickEventName,
 	dispatchClickEvent,
