+++ app/modules/onboarding/assets/js/utils/modules/onboarding-tracker.js
@@ -302,21 +302,6 @@ class OnboardingTracker {
 		}
 	}
 
-	sendThemeChoiceEvent( currentStep, themeValue ) {
-		if ( EventDispatcher.canSendEvents() ) {
-			return EventDispatcher.dispatchStepEvent(
-				ONBOARDING_EVENTS_MAP.THEME_CHOICE,
-				2,
-				ONBOARDING_STEP_NAMES.HELLO_BIZ,
-				{
-					location: 'plugin_onboarding',
-					trigger: 'theme_selected',
-					theme: themeValue,
-				},
-			);
-		}
-	}
-
 	sendTopUpgrade( currentStep, upgradeClicked ) {
 		return this.sendEventOrStore( 'TOP_UPGRADE', { currentStep, upgradeClicked } );
 	}
@@ -326,7 +311,6 @@ class OnboardingTracker {
 	}
 
 	initiateCoreOnboarding() {
-		StorageManager.clearAllOnboardingData();
 		TimingManager.clearStaleSessionData();
 		TimingManager.initializeOnboardingStartTime();
 	}
@@ -835,17 +819,14 @@ class OnboardingTracker {
 	}
 
 	onStepLoad( currentStep ) {
-		const stepNumber = this.getStepNumber( currentStep );
-		TimingManager.trackStepStartTime( stepNumber );
+		TimingManager.trackStepStartTime( this.getStepNumber( currentStep ) );
 
-		if ( 2 === stepNumber || 'hello' === currentStep || 'hello_biz' === currentStep ) {
+		if ( 2 === this.getStepNumber( currentStep ) || 'hello' === currentStep || 'hello_biz' === currentStep ) {
 			this.sendStoredStep1EventsOnStep2();
-			this.sendThemeSelectionExperimentStarted();
 		}
 
-		if ( 4 === stepNumber || 'goodToGo' === currentStep ) {
+		if ( 4 === this.getStepNumber( currentStep ) || 'goodToGo' === currentStep ) {
 			this.checkAndSendReturnToStep4();
-			this.sendGoodToGoExperimentStarted();
 		}
 	}
 
@@ -870,108 +851,6 @@ class OnboardingTracker {
 
 	clearAllOnboardingStorage() {
 		return PostOnboardingTracker.clearAllOnboardingStorage();
-	}
-
-	isThemeSelectionExperimentEnabled() {
-		return elementorAppConfig?.onboarding?.themeSelectionExperimentEnabled || false;
-	}
-
-	isGoodToGoExperimentEnabled() {
-		return elementorAppConfig?.onboarding?.goodToGoExperimentEnabled || false;
-	}
-
-	getThemeSelectionVariant() {
-		const stored = StorageManager.getString( ONBOARDING_STORAGE_KEYS.THEME_SELECTION_VARIANT );
-		if ( stored ) {
-			return stored;
-		}
-
-		return null;
-	}
-
-	getGoodToGoVariant() {
-		const stored = StorageManager.getString( ONBOARDING_STORAGE_KEYS.GOOD_TO_GO_VARIANT );
-		if ( stored ) {
-			return stored;
-		}
-
-		return null;
-	}
-
-	assignThemeSelectionVariant() {
-		if ( ! this.isThemeSelectionExperimentEnabled() ) {
-			return null;
-		}
-
-		const variant = Math.random() < 0.5 ? 'A' : 'B';
-		StorageManager.setString( ONBOARDING_STORAGE_KEYS.THEME_SELECTION_VARIANT, variant );
-		return variant;
-	}
-
-	assignGoodToGoVariant() {
-		if ( ! this.isGoodToGoExperimentEnabled() ) {
-			return null;
-		}
-
-		const variant = Math.random() < 0.5 ? 'A' : 'B';
-		StorageManager.setString( ONBOARDING_STORAGE_KEYS.GOOD_TO_GO_VARIANT, variant );
-		return variant;
-	}
-
-	sendThemeSelectionExperimentStarted() {
-		if ( StorageManager.exists( ONBOARDING_STORAGE_KEYS.THEME_SELECTION_EXPERIMENT_STARTED ) ) {
-			return;
-		}
-
-		let variant = this.getThemeSelectionVariant();
-
-		if ( ! variant ) {
-			variant = this.assignThemeSelectionVariant();
-			if ( ! variant ) {
-				return;
-			}
-		}
-
-		if ( ! EventDispatcher.canSendEvents() ) {
-			return;
-		}
-
-		const eventData = {
-			'Experiment name': 'core_onboarding_theme_selection',
-			'Variant name': variant,
-		};
-
-		EventDispatcher.dispatch( '$experiment_started', eventData );
-
-		StorageManager.setString( ONBOARDING_STORAGE_KEYS.THEME_SELECTION_EXPERIMENT_STARTED, 'true' );
-	}
-
-	sendGoodToGoExperimentStarted() {
-		if ( StorageManager.exists( ONBOARDING_STORAGE_KEYS.GOOD_TO_GO_EXPERIMENT_STARTED ) ) {
-			return;
-		}
-
-		let variant = this.getGoodToGoVariant();
-
-		if ( ! variant ) {
-			variant = this.assignGoodToGoVariant();
-			if ( ! variant ) {
-				return;
-			}
-		}
-
-		if ( ! EventDispatcher.canSendEvents() ) {
-			return;
-		}
-
-		const eventData = {
-			'Experiment name': 'core_onboarding_good_to_go',
-			'Variant name': variant,
-		};
-
-		EventDispatcher.dispatch( '$experiment_started', eventData );
-
-		StorageManager.setString( ONBOARDING_STORAGE_KEYS.GOOD_TO_GO_EXPERIMENT_STARTED, 'true' );
 	}
 }
