+++ app/modules/onboarding/assets/js/utils/modules/post-onboarding-tracker.js
@@ -58,7 +58,6 @@ class PostOnboardingTracker {
 
 	clearAllOnboardingStorage() {
 		StorageManager.clearAllOnboardingData();
-		StorageManager.clearExperimentData();
 	}
 }
